feat(auth): set up authentication and authorization
- implement endpoint for user signup and login
- implement middleware for authenticating jwt tokens
- implement role checking middleware

[Delivers #2]
