+++ __tests__/test.test.ts
@@ -1,20 +1,11 @@
-const { tester } = require('../src/controllers/testController');
+export {}
+const request = require('supertest');
+const app = require('../src/app');
 
-const res: any = {};
-
-(res.json = jest.fn((x: Object) => x)),
-  (res.status = jest.fn((x: number) => res));
-
-const req: any = {
-  body: {
-    name: 'test',
-  },
-};
 
 describe('Test', () => {
   it('should return 200 successful upon testing route', async () => {
-    await tester(req, res);
-
-    expect(res.status).toHaveBeenCalledWith(200);
+    const response = await request(app).get('/api/test')
+    expect(response.status).toBe(200);
   });
 });
